build: update quote exports
